Move 2 dao classes
