+++ corrigan/corrigan-impl/src/main/java/com/ait/corrigan/services/PayServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ait.corrigan.services;
 
+import com.ait.corrigan.dao.OrderDaoImpl;
 import com.ait.corrigan.dao.OrderDao;
 import com.ait.corrigan.dao.PaymentDao;
 import com.ait.corrigan.dao.PaymentDaoImpl;
