+++ corrigan/corrigan-impl/src/test/java/com/ait/corrigan/dao/OrderDaoImplIT.java
@@ -5,7 +5,6 @@
  */
 package com.ait.corrigan.dao;
 
-import com.ait.corrigan.services.OrderDaoImpl;
 import org.junit.Before;
 
 /**
